Rename styles file
